Refactor menu internals (#1867)
* Keymaps part@1 move menus
* Gather all paths in a module
