+++ app/menus/menus/plugins.js
@@ -1,0 +1,16 @@
+const {accelerators} = require('../../accelerators');
+
+module.exports = function (update) {
+  return {
+    label: 'Plugins',
+    submenu: [
+      {
+        label: 'Update',
+        accelerator: accelerators.updatePlugins,
+        click() {
+          update();
+        }
+      }
+    ]
+  };
+};
